subserver/pyserver/pyproject.toml
[project]
name = "xrk-agt-pyserver"
version = "1.0.0"
description = "XRK-AGT Python 子服务端"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "pyyaml>=6.0.1",
  "sentence-transformers>=5.2.0",
  "chromadb>=1.4.0",
  "langchain>=0.3.0",
  "langchain-openai>=0.2.0",
  "langgraph>=0.2.0",
]

[project.scripts]
xrk = "main:main"

[tool.uv]
package = false

[tool.setuptools]
py-modules = ["main"]
packages = ["apis", "core"]
